requirements.txt
streamlit
requests
python-dotenv
ibm-watsonx-orchestrate
